+++ src/Breadcrumbs/Breadcrumbs.spec.tsx
@@ -1,25 +1,14 @@
 import { render } from "@testing-library/react";
-import { Breadcrumbs } from "./Breadcrumbs";
+import { Breadcrumbs, BreadcrumbItem } from "./Breadcrumbs";
 
 describe("Breadcrumbs", () => {
-  const testBreadcrumbs = [
-    {
-      text: "Home",
-      href: "/",
-    },
-    {
-      text: "Getting started",
-      href: "/getting-started",
-    },
-    {
-      text: "Installation",
-      href: "/getting-started/installation",
-    },
-  ];
-
   it('should render a nav with an aria-label of "breadcrumbs"', () => {
     const { getByLabelText } = render(
-      <Breadcrumbs breadcrumbs={testBreadcrumbs} />
+      <Breadcrumbs>
+        <BreadcrumbItem href="/">Home</BreadcrumbItem>
+        <BreadcrumbItem href="/getting-started">Getting started</BreadcrumbItem>
+        <BreadcrumbItem isCurrentPage>Installation</BreadcrumbItem>
+      </Breadcrumbs>
     );
     expect(getByLabelText("breadcrumbs")).toBeTruthy();
     expect(getByLabelText("breadcrumbs").tagName).toBe("NAV");
@@ -27,25 +16,74 @@ describe("Breadcrumbs", () => {
 
   it("should render the provided aria-label", () => {
     const { getByLabelText } = render(
-      <Breadcrumbs breadcrumbs={testBreadcrumbs} aria-label="custom" />
+      <Breadcrumbs aria-label="custom">
+        <BreadcrumbItem href="/">Home</BreadcrumbItem>
+        <BreadcrumbItem href="/getting-started">Getting started</BreadcrumbItem>
+        <BreadcrumbItem isCurrentPage>Installation</BreadcrumbItem>
+      </Breadcrumbs>
     );
     expect(getByLabelText("custom")).toBeTruthy();
   });
 
   it("should render the provided breadcrumbs", () => {
-    const { getByText } = render(<Breadcrumbs breadcrumbs={testBreadcrumbs} />);
+    const { getByText } = render(
+      <Breadcrumbs>
+        <BreadcrumbItem href="/">Home</BreadcrumbItem>
+        <BreadcrumbItem href="/getting-started">Getting started</BreadcrumbItem>
+        <BreadcrumbItem isCurrentPage>Installation</BreadcrumbItem>
+      </Breadcrumbs>
+    );
     expect(getByText("Home")).toBeTruthy();
     expect(getByText("Getting started")).toBeTruthy();
     expect(getByText("Installation")).toBeTruthy();
   });
 
-  it("should render the last breaddcrumb with aria-current='page'", () => {
-    const { getByText } = render(<Breadcrumbs breadcrumbs={testBreadcrumbs} />);
+  it("should render the last breadcrumb with aria-current='page'", () => {
+    const { getByText } = render(
+      <Breadcrumbs>
+        <BreadcrumbItem href="/">Home</BreadcrumbItem>
+        <BreadcrumbItem href="/getting-started">Getting started</BreadcrumbItem>
+        <BreadcrumbItem isCurrentPage>Installation</BreadcrumbItem>
+      </Breadcrumbs>
+    );
     expect(getByText("Installation").getAttribute("aria-current")).toBe("page");
   });
 
-  it("should not render the last breadcrumb as a link", () => {
-    const { getByText } = render(<Breadcrumbs breadcrumbs={testBreadcrumbs} />);
-    expect(getByText("Installation").tagName).toBe("LI");
+  it("should render the separator between breadcrumbs", () => {
+    const { container } = render(
+      <Breadcrumbs>
+        <BreadcrumbItem href="/">Home</BreadcrumbItem>
+        <BreadcrumbItem href="/getting-started">Getting started</BreadcrumbItem>
+        <BreadcrumbItem isCurrentPage>Installation</BreadcrumbItem>
+      </Breadcrumbs>
+    );
+    // separator icons from react-lucide have lucide class
+    expect(container.querySelectorAll('.lucide')).toHaveLength(2);
+  });
+
+  it("should only render link if there is an href", () => {
+    const { getByText } = render(
+      <Breadcrumbs>
+        <BreadcrumbItem href="/">Home</BreadcrumbItem>
+        <BreadcrumbItem href="/getting-started">Getting started</BreadcrumbItem>
+        <BreadcrumbItem isCurrentPage>Installation</BreadcrumbItem>
+      </Breadcrumbs>
+    );
+    expect(getByText("Home").tagName).toBe("A");
+    expect(getByText("Installation").tagName).not.toBe("A");
+  });
+
+  it("should warn if BreadcrumbItem has more than one child when using asChild prop", () => {
+    const spy = jest.spyOn(console, "warn").mockImplementation();
+    render(
+      <Breadcrumbs>
+        <BreadcrumbItem asChild>
+          <a href="/">Home</a>
+          <span>Home</span>
+        </BreadcrumbItem>
+      </Breadcrumbs>
+    );
+    expect(spy).toHaveBeenCalled();
+    spy.mockRestore();
   });
 });